Add index=pas to setup.xml department search
Cloud demo users do not install and setup via the readme, so the steps to add a given username to the pasuser group are not shown. Since you're not initially a member, the search as written without an index will not include the pas index, so the searchbox to select departments for the dashboard won't work.

Adding the index to the search resolves the problem more easily than having users add themselves to a group, making it easier for new users to consume.